Initialize general domain_worker.rs by extracting the common logic from core_domain_worker.rs
`domain_worker.rs` is initialized from core_domain_worker.rs, it will be
evolved into a generic domain worker used by both the core_domain_worker
and system_domain_worker in the next commit.
