+++ domains/client/domain-executor/src/lib.rs
@@ -69,6 +69,7 @@ mod core_executor;
 mod domain_block_processor;
 mod domain_bundle_producer;
 mod domain_bundle_proposer;
+mod domain_worker;
 mod fraud_proof;
 mod merkle_tree;
 mod system_bundle_processor;
